+++ cypress/component/rich-text/RichTextEditor.Commands.spec.ts
@@ -64,7 +64,7 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
       //this is used instead of snapshot value because we have randomized entry IDs
       richText.getValue().should((doc) => {
         expect(
-          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY)
+          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY),
         ).to.have.length(1);
       });
     });
@@ -77,7 +77,7 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
       //this is used instead of snapshot value because we have randomized entry IDs
       richText.getValue().should((doc) => {
         expect(
-          doc.content[0].content.filter((node) => node.nodeType === INLINES.EMBEDDED_ENTRY)
+          doc.content[0].content.filter((node) => node.nodeType === INLINES.EMBEDDED_ENTRY),
         ).to.have.length(1);
       });
     });
@@ -92,7 +92,7 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
         block(BLOCKS.EMBEDDED_ASSET, {
           target: { sys: { id: 'published_asset', type: 'Link', linkType: 'Asset' } },
         }),
-        block(BLOCKS.PARAGRAPH, {}, text())
+        block(BLOCKS.PARAGRAPH, {}, text()),
       );
 
       richText.expectValue(expectedValue);
@@ -107,42 +107,45 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
     });
 
     it('should navigate then embed on pressing enter', () => {
+      richText.editor.click().type('testing enter{enter}');
       richText.editor.click().type('/');
       getCommandList().findByText('Embed Example Content Type').should('exist');
       richText.editor.type('{enter}');
       getCommandList().findByText('Embed Example Content Type').should('not.exist');
       richText.editor.type('{enter}');
-
       //this is used instead of snapshot value because we have randomized entry IDs
       richText.getValue().should((doc) => {
         expect(
-          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY)
+          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY),
         ).to.have.length(1);
       });
     });
 
     it('should select next item on down arrow press', () => {
-      richText.editor.click().type('/{downarrow}{enter}{enter}');
+      richText.editor.click().type('/');
+      getCommandList().findByText('Embed Example Content Type').should('exist');
+      richText.editor.type('{enter}');
+      richText.editor.type('{enter}');
+      richText.editor.type('{downarrow}');
 
-      richText.editor.findByTestId('embedded-entry-inline').should('exist');
+      cy.findByRole('article').should('exist');
 
-      //this is used instead of snapshot value because we have randomized entry IDs
-      richText.getValue().should((doc) => {
-        expect(
-          doc.content[0].content.filter((node) => node.nodeType === INLINES.EMBEDDED_ENTRY)
-        ).to.have.length(1);
-      });
+      //check for the CSS style that indicates selection
+      cy.findByRole('article').should('have.css', 'border-color', 'rgb(3, 111, 227)'); // tokens.blue300
     });
 
     it('should select previous item on up arrow press', () => {
-      richText.editor.click().type('/{downarrow}{uparrow}{enter}{enter}');
+      richText.editor.click().type('testing preview item.');
+      richText.editor.click().type('/');
+      getCommandList().findByText('Embed Example Content Type').should('exist');
+      richText.editor.type('{enter}');
+      richText.editor.type('{enter}');
+      richText.editor.type('{downarrow}');
+      richText.editor.type('{uparrow}');
+      cy.findByRole('article').should('exist');
 
-      //this is used instead of snapshot value because we have randomized entry IDs
-      richText.getValue().should((doc) => {
-        expect(
-          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY)
-        ).to.have.length(1);
-      });
+      //check for the CSS style that indicates it is not selected
+      cy.findByRole('article').should('have.css', 'border-color', 'rgb(207, 217, 224)');
     });
 
     it('should not delete adjacent text', () => {
@@ -152,7 +155,7 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
       richText.getValue().should((doc) => {
         expect(doc.content[0].content[0].value).to.equal('test');
         expect(
-          doc.content[0].content.filter((node) => node.nodeType === INLINES.EMBEDDED_ENTRY)
+          doc.content.filter((node) => node.nodeType === BLOCKS.EMBEDDED_ENTRY),
         ).to.have.length(1);
       });
     });
@@ -167,7 +170,7 @@ describe('Rich Text Editor - Commands', { viewportHeight: 2000, viewportWidth: 1
         expect(
           doc.content.filter((node) => {
             return node.nodeType === BLOCKS.EMBEDDED_ENTRY || node.nodeType === BLOCKS.HEADING_1;
-          })
+          }),
         ).to.have.length(2);
       });
     });