+++ packages/_shared/src/index.tsx
@@ -42,4 +42,4 @@ export * from './ReleaseEntityStatusBadge';
 export * from './utils/determineReleaseAction';
 export * from './utils/getEntityReleaseStatus';
 export * from './utils/getReleaseStatusBadgeConfig';
-export { SharedQueryClientProvider, useQueryClient, useQuery } from './queryClient';
+export * from './queryKeys';