+++ packages/rich-text-alpha/src/components/FetchingWrappedEntryCard.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { FieldAppSDK } from '@contentful/app-sdk';
+import { ContentType, FieldAppSDK } from '@contentful/app-sdk';
 import { ScheduledAction, Entry } from '@contentful/app-sdk';
 import { DragHandle, EntryCard } from '@contentful/f36-components';
 import {
@@ -18,6 +18,10 @@ import {
   type ReleaseV2Props,
   type ReleaseEntityStatus,
 } from '@contentful/field-editor-shared';
+import {
+  SharedQueryClientProvider,
+  useContentType,
+} from '@contentful/field-editor-shared/react-query';
 import areEqual from 'fast-deep-equal';
 
 interface InternalEntryCard {
@@ -27,6 +31,7 @@ interface InternalEntryCard {
   sdk: FieldAppSDK;
   loadEntityScheduledActions: (entityType: string, entityId: string) => Promise<ScheduledAction[]>;
   entry: Entry;
+  contentType?: ContentType;
   onEdit?: VoidFunction;
   onRemove?: VoidFunction;
   localesStatusMap?: LocalePublishStatusMap;
@@ -46,13 +51,11 @@ const InternalEntryCard = React.memo(
     isSelected,
     isDisabled,
     locale,
+    contentType,
     onEdit,
     onRemove,
     localesStatusMap,
   }: InternalEntryCard) => {
-    const contentType = sdk.space
-      .getCachedContentTypes()
-      .find((contentType) => contentType.sys.id === entry.sys.contentType.sys.id);
     const activeLocales = useActiveLocales(sdk);
 
     return (
@@ -99,7 +102,7 @@ interface FetchingWrappedEntryCardProps {
   onRemove?: VoidFunction;
 }
 
-export const FetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) => {
+const InternalFetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) => {
   const { entryId, onEntityFetchComplete } = props;
   const { data: entry, status, currentEntity } = useEntity<Entry>('Entry', entryId);
   const { getEntityScheduledActions } = useEntityLoader();
@@ -114,6 +117,11 @@ export const FetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) =
     locales: props.sdk.locales,
     release: props.sdk.release,
     isReference: true,
+  });
+
+  const contentTypeId = entry?.sys.contentType.sys.id;
+  const { data: contentType } = useContentType(props.sdk, contentTypeId || '', {
+    enabled: !!contentTypeId,
   });
 
   React.useEffect(() => {
@@ -141,6 +149,7 @@ export const FetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) =
       entry={entry}
       sdk={props.sdk}
       locale={props.locale}
+      contentType={contentType}
       isDisabled={props.isDisabled}
       isSelected={props.isSelected}
       onEdit={props.onEdit}
@@ -151,5 +160,13 @@ export const FetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) =
       release={props.sdk.release as ReleaseV2Props | undefined}
       releaseEntityStatus={releaseEntityStatus}
     />
+  );
+};
+
+export const FetchingWrappedEntryCard = (props: FetchingWrappedEntryCardProps) => {
+  return (
+    <SharedQueryClientProvider>
+      <InternalFetchingWrappedEntryCard {...props} />
+    </SharedQueryClientProvider>
   );
 };