+++ packages/rich-text/src/internal/index.ts
@@ -2,3 +2,4 @@ export * from './queries';
 export * from './types';
 export * from './misc';
 export * from './transforms';
+export * from './selection';