+++ packages/rich-text-alpha/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.7](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.6...@contentful/field-editor-rich-text-alpha@0.4.7) (2026-02-27)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.6](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.5...@contentful/field-editor-rich-text-alpha@0.4.6) (2026-02-23)
+
+### Bug Fixes
+
+- **shared:** make React Query optional + ESM/Vite-safe via subpath export [TOL-38/46] ([#2079](https://github.com/contentful/field-editors/issues/2079)) ([64021a2](https://github.com/contentful/field-editors/commit/64021a263371c3aade8dcab93b727a012a1bce30))
+
+## [0.4.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.4...@contentful/field-editor-rich-text-alpha@0.4.5) (2026-01-29)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.3...@contentful/field-editor-rich-text-alpha@0.4.4) (2026-01-29)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.2...@contentful/field-editor-rich-text-alpha@0.4.3) (2026-01-28)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.1...@contentful/field-editor-rich-text-alpha@0.4.2) (2026-01-28)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+## [0.4.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.4.0...@contentful/field-editor-rich-text-alpha@0.4.1) (2026-01-23)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
+# [0.4.0](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.3.4...@contentful/field-editor-rich-text-alpha@0.4.0) (2026-01-22)
+
+### Features
+
+- Migrate field editors .space.getCachedContentTypes to Plain Client APIs [TOL-3472] ([#2043](https://github.com/contentful/field-editors/issues/2043)) ([4f9ab95](https://github.com/contentful/field-editors/commit/4f9ab952b7b70b900a00b28695358d3b8d722b2e))
+
+## [0.3.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.3.3...@contentful/field-editor-rich-text-alpha@0.3.4) (2026-01-21)
+
+**Note:** Version bump only for package @contentful/field-editor-rich-text-alpha
+
 ## [0.3.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rich-text-alpha@0.3.2...@contentful/field-editor-rich-text-alpha@0.3.3) (2026-01-19)
 
 **Note:** Version bump only for package @contentful/field-editor-rich-text-alpha